requirements.txt
flask==3.1.0
flask-cors==5.0.0
bcrypt==4.2.1
python-dotenv==1.0.1
pytz==2024.2
Werkzeug==3.1.3
gunicorn==23.0.0
Pillow==12.1.1
requests==2.32.3
reportlab==4.2.5
boto3==1.35.99
psycopg2-binary==2.9.10
